add ColorButton class and constants; remove obsolete color button files
