--- code/color_memory/constants.py
@@ -1,17 +1,0 @@
-RED = (255,0,0)
-LIGHT_RED = (255,100,100)
-BLUE = (0,0,255)
-LIGHT_BLUE = (100,100,255)
-GREEN = (0,255,0)
-LIGHT_GREEN = (100,255,100)
-YELLOW = (255,180,0)
-LIGHT_YELLOW = (255,255,50)
-WHITE = (255,255,255)
-BLACK = (0,0,0)
-SCREEN_WIDTH = 720
-SCREEN_HEIGHT = 720
-
-PORT_GREEN = 22
-PORT_RED = 12
-PORT_YELLOW = 27
-PORT_BLUE = 17
