+++ code/color_memory/led_buttons.py
@@ -4,11 +4,6 @@ try:
     import RPi.GPIO as GPIO
     from rpi_ws281x import PixelStrip, Color
 except ImportError:
-    import sys
-    import os
-
-    sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-
     from mock_gpio import MockGPIO as GPIO
     from mock_ws281x import PixelStrip, Color
